Autenticação sendo enviada junto da requisição, para que liberer acesso à base de dados. Um usuário só pode alterar ou excluir se for o mesmo usuário logado

+++ src/controllers/UsuarioController.php
@@ -10,14 +10,31 @@ class UsuarioController extends Controller {
     //Listando todos os usuários.
     public function index(){
 
-        //$array = array('erro' => '');
+        $array = array('loged' => false);
 
         $method = parent::getMethodRequisition();
+        $data = parent::getRequestData();
 
-        if($method == 'GET'){
+        if(!empty($data['jwt'])){
 
-            $array['usuarios'] = UsuarioDao::getAllUser();
+            if($array['logedUserId'] = UsuarioDao::validateJwt($data['jwt'])){
+                $array['loged'] = true;
 
+                if($method == 'GET'){
+                    $array['usuarios'] = UsuarioDao::getAllUser();
+                }else{
+                    parent::setResponseStatus(404);
+                    $array['error'] = "Método indisponível";
+                }
+
+            }else{
+                parent::setResponseStatus(401);
+                $array['error'] = "Acesso negado";
+            }
+
+        }else{
+            parent::setResponseStatus(401);
+            $array['error'] = "Acesso negado";
         }
 
         parent::returnJson($array);
@@ -39,18 +56,20 @@ class UsuarioController extends Controller {
                 $usuario->setSenha(addslashes($data['senha']));
 
                 if(UsuarioDao::login($usuario)){
-                    //gerar JTW
-                    //$array['logedUser'] = $usuario->getLogedUser();
+                    //Gerando JTW para o id do usuário logado
                     $idLogedUser = $usuario->getLogedUser();
                     $array['jwt'] = UsuarioDao::createJwt($idLogedUser);
                 }else{
+                    parent::setResponseStatus(403);
                     $array['error'] = "Acesso negado";
                 }
             }else{
+                parent::setResponseStatus(204);
                 $array['error'] = "Preencha os campos e-mail e senha";
             }
 
         }else{
+            parent::setResponseStatus(400);
             $array['error'] = "Método indisponível";
         }
 
@@ -72,24 +91,26 @@ class UsuarioController extends Controller {
 
                 $array['loged'] = true;
 
+                $array['isMe'] = UsuarioDao::isMe($array['logedUserId'], $args['id']);
+
                 if($method == "GET"){
                     $usuarios = UsuarioDao::getUser($args['id']);
 
                     if($usuarios){
                         $array['usuario'] = $usuarios;
                     }else{
-                        http_response_code(404);
+                        parent::setResponseStatus(404);
                         $array['error'] = "Usuário não existe na base de dados";
                     }
                 }
 
             }else{
-                http_response_code(401);
+                parent::setResponseStatus(401);
                 $array['error'] = "Acesso negado";
             }
 
         }else{
-            http_response_code(401);
+            parent::setResponseStatus(401);
             $array['error'] = "Acesso negado";
         }
 
@@ -118,21 +139,26 @@ class UsuarioController extends Controller {
 
                     if(UsuarioDao::insertUser($usuario)){
                         $array['usuario'] = $usuario->getId();
-                        $array['jwt'] = 'Token-JWT';
-                        http_response_code(201);
+
+                        parent::setResponseStatus(201);
+                        $array['success'] = 'Cadastro realizado';
                     }else{
+                        parent::setResponseStatus(203);
                         $array['error'] = "Este e-mail já existe na base de dados";
                     }
 
                 }else{
+                    parent::setResponseStatus(204);
                     $array['error'] = "O e-mail enviado não é válido";
                 }
 
             }else{
+                parent::setResponseStatus(204);
                 $array['error'] = "Ops! Algum campo não foi preenchido";
             }
 
         }else{
+            parent::setResponseStatus(400);
             $array['error'] = "Método de requisição indisponível";
         }
 
@@ -144,39 +170,70 @@ class UsuarioController extends Controller {
     public function updateUser($args = array()){
 
         $mudar = array();
-        $array = array();
+        $array = array('loged' => false);
 
         $mudar['id'] = $args['id'];
 
         $metodo = parent::getMethodRequisition();
 
-        if($metodo == "PUT"){
+        $data = parent::getRequestData();
 
-            $data = parent::getRequestData();
+        if(!empty($data['jwt'])){
 
-            if(!empty($data['nome'])){
-                $mudar['nome'] = $data['nome'];
-            }
-            if(!empty($data['email'])){
-                if(filter_var($data['email'], FILTER_VALIDATE_EMAIL)){
-                    if(UsuarioDao::emailValidation($data['email'])){
-                        $mudar['email'] = addslashes($data['email']);
+            if($metodo == "PUT"){
+
+                if($array['logedUserId'] = UsuarioDao::validateJwt($data['jwt'])){
+
+                    $array['loged'] = true;
+
+                    $array['isMe'] = UsuarioDao::isMe($array['logedUserId'], $args['id']);
+
+                    if($array['isMe'] == true){
+
+                        if(!empty($data['nome'])){
+                            $mudar['nome'] = $data['nome'];
+                        }
+                        if(!empty($data['email'])){
+                            if(filter_var($data['email'], FILTER_VALIDATE_EMAIL)){
+                                if(UsuarioDao::emailValidation($data['email'])){
+                                    $mudar['email'] = addslashes($data['email']);
+                                }else{
+                                    $array['error'] = "Este e-mail já existe na base de dados";
+                                }
+                            }else{
+                                $array['error'] = "Digite um e-mail válido";
+                            }
+                        }
+                        if(!empty($data['senha'])){
+                            $mudar['senha'] = password_hash(addslashes($data['senha']), PASSWORD_DEFAULT);
+                        }
+
+                        if(UsuarioDao::updateUser($args['id'], $mudar)){
+
+                            $array['update'] = "Usuario alterado com sucesso";
+                        }else{
+                            parent::setResponseStatus(404);
+                            $array['error'] = "Usuário não existe na base de dados";
+                        }
+
                     }else{
-                        $array['error'] = "Este e-mail já existe na base de dados";
+                        parent::setResponseStatus(401);
+                        $array['error'] = "Você não tem permissão para alterar este usuário.";
                     }
+
                 }else{
-                    $array['error'] = "Digite um e-mail válido";
+                    parent::setResponseStatus(401);
+                    $array['error'] = "Acesso negado";
                 }
-            }
-            if(!empty($data['senha'])){
-                $mudar['senha'] = password_hash(addslashes($data['senha']), PASSWORD_DEFAULT);
+
+            }else{
+                parent::setResponseStatus(404);
+                $array['error'] = "Método indisponível";
             }
 
-            if(UsuarioDao::updateUser($args['id'], $mudar)){
-                $array['update'] = "Usuario alterado com sucesso";
-            }else{
-                $array['error'] = "Usuário não existe na base de dados";
-            }
+        }else{
+            parent::setResponseStatus(401);
+            $array['error'] = "Acesso negado";
         }
 
         parent::returnJson($array);
@@ -186,19 +243,42 @@ class UsuarioController extends Controller {
     //Deletar usuario
     public function deleteUser($args = array()){
 
-        $array = array();
+        $array = array('loged' => false);
 
         $metodo = parent::getMethodRequisition();
+        $data = parent::getRequestData();
 
-        if($metodo == 'DELETE'){
+        if(!empty($data['jwt']) && $metodo == "DELETE"){
 
-            if(UsuarioDao::deleteUser($args['id']) == true){
-                $array['id'] = $args['id'];
-                $array['delete'] = "Usuário excluído com sucesso";
+            if($array['logedUserId'] = UsuarioDao::validateJwt($data['jwt'])){
+
+                $array['loged'] = true;
+
+                $array['isMe'] = UsuarioDao::isMe($array['logedUserId'], $args['id']);
+
+                if($array['isMe'] == true){
+
+                    if(UsuarioDao::deleteUser($args['id']) == true){
+                        $array['id'] = $args['id'];
+                        $array['delete'] = "Usuário excluído com sucesso";
+                    }else{
+                        parent::setResponseStatus(404);
+                        $array['error'] = "Usuário não existe na base de dados";
+                    }
+
+                }else{
+                    parent::setResponseStatus(401);
+                    $array['error'] = "Você não tem permissão para alterar este usuário.";
+                }
+
             }else{
-                $array['error'] = "Usuário não existe na base de dados";
+                parent::setResponseStatus(401);
+                $array['error'] = "Acesso negado";
             }
 
+        }else{
+            parent::setResponseStatus(401);
+            $array['error'] = "Método indisponível";
         }
 
         parent::returnJson($array);
